ABS.StdLib

ABS builtin types

The ABS standard datatypes, most coming from standard Haskell, except Fut coming from habs-runtime package

type Int = Int

type Rat = Ratio Int

data Bool :: *

Constructors

False

True

type Unit = ()

type List = []

An ABS synonym to underneath Haskell lists. It is the same as defining in ABS:

data List a = Nil | Cons a (List a)

type String = [Char]

A String is a list of characters. String constants in Haskell are values of type String.

Operations on numbers

(+) :: Num a => a -> a -> a

(-) :: Num a => a -> a -> a

(*) :: Num a => a -> a -> a

(%) :: Rat -> Rat -> Int

Modulo operation. Takes two Rats and returns an integer.

Truncated towards 0, so it is Haskell's rem.

abs :: Num a => a -> a

Absolute value.

pow :: Num a => a -> Int -> a

Raising a number to a non-negative integer power

Note: deviation, abstools defines pow also for negative integral powers, but then the result always will be a Rat: int/rat conversion issue

Another way is to always return a Rat with (^^) and then to explicitly truncate if you want Int.

Operations on rationals

numerator :: Ratio a -> a

Extract the numerator of the ratio in reduced form: the numerator and denominator have no common factor and the denominator is positive.

denominator :: Ratio a -> a

Extract the denominator of the ratio in reduced form: the numerator and denominator have no common factor and the denominator is positive.

Rational division. Takes any number but always returns a Rat.

(/) :: Fractional a => a -> a -> a

fractional division

Rat to Int conversion

truncate :: RealFrac a => forall b. Integral b => a -> b

truncate x returns the integer nearest x between zero and x

Boolean Operations

(||) :: Bool -> Bool -> Bool infixr 2

Boolean "or"

(&&) :: Bool -> Bool -> Bool infixr 3

Boolean "and"

(==) :: Eq a => a -> a -> Bool

not :: Bool -> Bool

Boolean "not"

Ordering operations

(<) :: Ord a => a -> a -> Bool

(<=) :: Ord a => a -> a -> Bool

(>=) :: Ord a => a -> a -> Bool

(>) :: Ord a => a -> a -> Bool

min :: Ord a => a -> a -> a

max :: Ord a => a -> a -> a

minimum :: Foldable t => forall a. Ord a => t a -> a

The least element of a non-empty structure.

maximum :: Foldable t => forall a. Ord a => t a -> a

The largest element of a non-empty structure.

Built-in Pairs and Triples and their functions

type Pair a b = (a, b)

fst :: (a, b) -> a

Extract the first component of a pair.

snd :: (a, b) -> b

Extract the second component of a pair.

type Triple a b c = (a, b, c)

fstT :: Triple a b c -> a

sndT :: Triple a b c -> b

trd :: Triple a b c -> c

fmap'Pair :: (a -> c) -> (b -> d) -> Pair a b -> Pair c d

fmap'Triple :: (a -> a1) -> (b -> b1) -> (c -> c1) -> (a, b, c) -> (a1, b1, c1)

Maybe, Either datatypes and their functions

data Maybe a :: * -> *

The Maybe type encapsulates an optional value. A value of type Maybe a either contains a value of type a (represented as Just a), or it is empty (represented as Nothing). Using Maybe is a good way to deal with errors or exceptional cases without resorting to drastic measures such as error.

The Maybe type is also a monad. It is a simple kind of error monad, where all errors are represented by Nothing. A richer error monad can be built using the Either type.

Constructors

Nothing

Just a

fromJust :: Maybe a -> a

The fromJust function extracts the element out of a Just and throws an error if its argument is Nothing.

Examples

Basic usage:

>>> fromJust (Just 1)
1

>>> 2 * (fromJust (Just 10))
20

>>> 2 * (fromJust Nothing)
*** Exception: Maybe.fromJust: Nothing

isJust :: Maybe a -> Bool

The isJust function returns True iff its argument is of the form Just _.

Examples

Basic usage:

>>> isJust (Just 3)
True

>>> isJust (Just ())
True

>>> isJust Nothing
False

Only the outer constructor is taken into consideration:

>>> isJust (Just Nothing)
True

fmap'Maybe :: (a -> b) -> Maybe a -> Maybe b

data Either a b :: * -> * -> *

The Either type represents values with two possibilities: a value of type Either a b is either Left a or Right b.

The Either type is sometimes used to represent a value which is either correct or an error; by convention, the Left constructor is used to hold an error value and the Right constructor is used to hold a correct value (mnemonic: "right" also means "correct").

Examples

The type Either String Int is the type of values which can be either a String or an Int. The Left constructor can be used only on Strings, and the Right constructor can be used only on Ints:

>>> let s = Left "foo" :: Either String Int
>>> s
Left "foo"
>>> let n = Right 3 :: Either String Int
>>> n
Right 3
>>> :type s
s :: Either String Int
>>> :type n
n :: Either String Int

The fmap from our Functor instance will ignore Left values, but will apply the supplied function to values contained in a Right:

>>> let s = Left "foo" :: Either String Int
>>> let n = Right 3 :: Either String Int
>>> fmap (*2) s
Left "foo"
>>> fmap (*2) n
Right 6

The Monad instance for Either allows us to chain together multiple actions which may fail, and fail overall if any of the individual steps failed. First we'll write a function that can either parse an Int from a Char, or fail.

>>> import Data.Char ( digitToInt, isDigit )
>>> :{
    let parseEither :: Char -> Either String Int
        parseEither c
          | isDigit c = Right (digitToInt c)
          | otherwise = Left "parse error"
>>> :}

The following should work, since both '1' and '2' can be parsed as Ints.

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither '1'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Right 3

But the following should fail overall, since the first operation where we attempt to parse 'm' as an Int will fail:

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither 'm'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Left "parse error"

Constructors

Left a

Right b

left :: Either a b -> a

Deconstructs _unsafely_ the left part of an Either

right :: Either a b -> b

Deconstructs _unsafely_ the right part of an Either

isLeft :: Either a b -> Bool

Return True if the given value is a Left-value, False otherwise.

Examples

Basic usage:

>>> isLeft (Left "foo")
True
>>> isLeft (Right 3)
False

Assuming a Left value signifies some sort of error, we can use isLeft to write a very simple error-reporting function that does absolutely nothing in the case of success, and outputs "ERROR" if any error occurred.

This example shows how isLeft might be used to avoid pattern matching when one does not care about the value contained in the constructor:

>>> import Control.Monad ( when )
>>> let report e = when (isLeft e) $ putStrLn "ERROR"
>>> report (Right 1)
>>> report (Left "parse error")
ERROR

Since: 4.7.0.0

isRight :: Either a b -> Bool

Return True if the given value is a Right-value, False otherwise.

Examples

Basic usage:

>>> isRight (Left "foo")
False
>>> isRight (Right 3)
True

Assuming a Left value signifies some sort of error, we can use isRight to write a very simple reporting function that only outputs "SUCCESS" when a computation has succeeded.

This example shows how isRight might be used to avoid pattern matching when one does not care about the value contained in the constructor:

>>> import Control.Monad ( when )
>>> let report e = when (isRight e) $ putStrLn "SUCCESS"
>>> report (Left "parse error")
>>> report (Right 1)
SUCCESS

Since: 4.7.0.0

fmap'Either :: (a -> c) -> (b -> d) -> Either a b -> Either c d

Functions for List datastructures

list :: [a] -> [a]

dummy function for ABS n-ary constructors

tail :: [a] -> [a]

Extract the elements after the head of a list, which must be non-empty.

head :: [a] -> a

Extract the first element of a list, which must be non-empty.

length :: Foldable t => forall a. t a -> Int

Returns the size/length of a finite structure as an Int. The default implementation is optimized for structures that are similar to cons-lists, because there is no general way to do better.

isEmpty :: List a -> Bool

Checks if the list is empty.

nth :: List a -> Int -> a

Returns the element of the list positioned at the given index.

concatenate :: [a] -> [a] -> [a]

appendright :: [a] -> a -> [a]

without :: Eq t => [t] -> t -> [t]

Removes all occurences of an element from a list

repeat :: a -> [a]

repeat x is an infinite list, with x the value of every element.

reverse :: [a] -> [a]

reverse xs returns the elements of xs in reverse order. xs must be finite.

copy :: a -> Int -> List a

Replicating an element n times, forming a list of length n.

fmap'List :: (a -> b) -> [a] -> [b]

The ABS Map datatype and its functions

data Map k a :: * -> * -> *

A Map from keys k to values a.

map :: Ord k => List (Pair k a) -> Map k a

Constructing maps from an association list.

_emptyMap :: Map k a

Constructor of empty Maps, EmptyMap in ABS

put :: Ord k => Map k v -> k -> v -> Map k v

insert :: Ord k => Map k v -> (k, v) -> Map k v

lookup :: Ord k => Map k v -> k -> Maybe v

lookupMaybe :: Ord k => Map k v -> k -> Maybe v

Deprecated: Use lookup instead

Synonym to lookup

lookupUnsafe :: Ord k => Map k v -> k -> v

lookupDefault :: Ord k => Map k a -> k -> a -> a

Returns the value associated with key k in map ms, or the value d if k has no entry in ms.

removeKey :: Ord k => Map k v -> k -> Map k v

keys :: Map k a -> Set k

values :: Map k a -> List a

The ABS Set datatype and its functions

data Set a :: * -> *

A set of values a.

set :: Ord a => List a -> Set a

_emptySet :: Set a

emptySet :: Set a -> Bool

size :: Set a -> Int

O(1). The number of elements in the set.

contains :: Ord a => Set a -> a -> Bool

union :: Ord a => Set a -> Set a -> Set a

O(n+m). The union of two sets, preferring the first set when equal elements are encountered. The implementation uses the efficient hedge-union algorithm.

intersection :: Ord a => Set a -> Set a -> Set a

O(n+m). The intersection of two sets. The implementation uses an efficient hedge algorithm comparable with hedge-union. Elements of the result come from the first set, so for example

import qualified Data.Set as S
data AB = A | B deriving Show
instance Ord AB where compare _ _ = EQ
instance Eq AB where _ == _ = True
main = print (S.singleton A `S.intersection` S.singleton B,
              S.singleton B `S.intersection` S.singleton A)

prints (fromList [A],fromList [B]).

difference :: Ord a => Set a -> Set a -> Set a

O(n+m). Difference of two sets. The implementation uses an efficient hedge algorithm comparable with hedge-union.

insertElement :: Ord a => Set a -> a -> Set a

remove :: Ord a => Set a -> a -> Set a

take :: Set a -> a

Returns one (arbitrary) element from a set. To iterate over a set, take one element and remove it from the set. Repeat until set is empty.

hasNext :: Set a -> Bool

next :: Set a -> (Set a, a)

NB: partial function, combine it with hasNext.

Printing to Strings and to standard-output

toString :: Show a => a -> String

intToString :: Int -> String

Returns a string with the base-10 textual representation of n. Note: Will work the same as toString. Just a carry-over from the other frontend.

substr :: String -> Int -> Int -> String

Returns a substring of string str of the given length starting from start (inclusive) Where the first character has index 0

Example: substr("abcde",1,3) => "bcd"

strlen :: String -> Int

Lifting ABS pure code to ABS object layer

Haskell is pure by default. These are necessary functions for lifting pure ABS expressions (of the functional core) to the ABS' object layer (monadic statements).

Haskell's return

is an expression taking a pure value and lifting it to the monadic world.

return :: Monad m => forall a. a -> m a

Inject a value into the monadic type.

Orphan instances

Num [Char]

Methods

(+) :: [Char] -> [Char] -> [Char]

(-) :: [Char] -> [Char] -> [Char]

(*) :: [Char] -> [Char] -> [Char]

negate :: [Char] -> [Char]

abs :: [Char] -> [Char]

signum :: [Char] -> [Char]

fromInteger :: Integer -> [Char]